requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-constance==2.9.1
django-js-asset==2.0.0
django-picklefield==3.1
django-taggit==3.0.0
djangorestframework==3.13.1
httplib2==0.20.4
install==1.3.5
oauth2client==4.1.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.2.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
